Safari-611.1.10/LayoutTests/webaudio/gain-basic-expected.txt

Tests GainNode attributes in IDL.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS gain is of AudioParam type.
PASS successfullyParsed is true

TEST COMPLETE
